REFACTOR: Moved group function to actions and changed drop down prop - Moved grouping function so that it can be used by other components - Changed prop type to function so that drop down component also works with things that do not use useState
